pyproject.toml
[project]
name = "g2pt"
version = "0.1.0"
description = "General Geometry Pretrained Transformer for 3D Point Cloud Processing"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "h5py>=3.15.1",
    "hydra-core==1.3.0",
    "liger-kernel==0.6.3",
    "lightning>=2.5.0",
    "loguru>=0.7.0",
    "matplotlib>=3.10.0",
    "meshio>=5.3.5",
    "mlflow>=3.5.0",
    "open3d>=0.19.0",
    "psutil>=7.0.0",
    "pyqt6>=6.9.1,<6.10",
    "rich>=14.0.0",
    "robust-laplacian>=1.0.0",
    "torch==2.6.0",
    "torch-geometric==2.7.0",
    "torchvision==0.21.0",
    "triton==3.2.0",
    "trimesh>=4.6.12",
    "torchmetrics>=1.8.2",
    "einops>=0.8.1",
    "opt-einsum>=3.4.0",
    "pytest>=9.0.1",
    "pyamg>=5.3.0",
]

[project.scripts]
g2pt = "g2pt:yzr"
[build-system]
requires = ["uv_build>=0.9.4,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
pg = [
    "torch-cluster",
]
pg_ext = [ # To accelerate point-transformer
    "torch-cluster",
    "torch-scatter",
]
pg_all = [ # Full Requirement of torch-geometric
    "torch-cluster",
    "torch-scatter",
    "pyg_lib",
    "torch-sparse",
    "torch-spline-conv",
]
dev = [
    "fast-simplification>=0.1.13",
    "ruff>=0.14.10",
    "tetgen>=0.7.0",
]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[[tool.uv.index]]
name = "pyg-cu126"
url = "https://data.pyg.org/whl/torch-2.6.0+cu126.html"
format = "flat"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu126" },
]
torchvision = [
  { index = "pytorch-cu126" },
]
torch-cluster = { index = "pyg-cu126" }
torch-scatter = { index = "pyg-cu126" }
torch-sparse = { index = "pyg-cu126" }
torch-spline-conv = { index = "pyg-cu126" }
pyg_lib = { index = "pyg-cu126" }

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 120
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`) codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"

[tool.basedpyright]
include = ["src"]
exclude = ["data", "data_exp",
    "**/__pycache__",
]
reportMissingTypeStubs = false
typeCheckingMode = 'basic'
